Dispatcher
* Dispatcher WIP

* Fixed recompilation of generated code even if nothing changed
Continue implementation of dispatcher

* Dispatcher WIP

* Fixed recompilation of generated code even if nothing changed
Continue implementation of dispatcher

* Do not rebuild codegen tool each time

* Make codegen tool run in release mode on configure

* Fixed serialization codegen
Implemented templated http requests

* Fixed codegen tool

* Implemented easy http requests

* Fixed test request on backend

* Fixed mac compilation

* Implemented request authentication

+++ CodegenTool/CodegenTool/OutputFilesManager.cs
@@ -34,16 +34,25 @@ public class OutputFilesManager
     {
         Directory.CreateDirectory(root);
         
-        foreach (var (file, content) in _files)
+        foreach (var (file, sb) in _files)
         {
             var path = Path.Combine(root, file);
+            var contents = sb.ToString();
 
             if (File.Exists(path))
             {
+                var oldContents = File.ReadAllText(path);
+
+                if (contents.Equals(oldContents, StringComparison.InvariantCulture))
+                {
+                    Console.WriteLine($"[INFO] {Path.GetFileName(path)} is unchanged. Skipping.");
+                    continue;
+                }
+                
                 File.Delete(path);
             }
 
-            File.WriteAllText(path, content.ToString());
+            File.WriteAllText(path, contents);
             File.SetLastWriteTime(path, DateTime.Now);
         }
     }
